requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
openai==1.1.1
pydantic==2.4.2
python-dotenv==1.0.1
icecream==2.1.3
sqlalchemy==2.0.27
asyncpg==0.29.0
anthropic==0.19.2
